split redirect tests into separate file
